requirements.txt
aiofiles==24.1.0
aiogram==3.13.1
aiohappyeyeballs==2.4.3
aiohttp==3.10.10
aiosignal==1.3.1
annotated-types==0.7.0
asgiref==3.8.1
attrs==24.2.0
certifi==2024.8.30
Django==5.1.2
django-cors-headers==4.5.0
djangorestframework==3.15.2
drf-spectacular==0.27.2
drf-yasg==1.21.7
frozenlist==1.4.1
idna==3.10
inflection==0.5.1
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
magic-filter==1.0.12
multidict==6.1.0
packaging==24.1
propcache==0.2.0
psycopg2==2.9.9
pydantic==2.9.2
pydantic_core==2.23.4
python-dotenv==1.0.1
pytz==2024.2
PyYAML==6.0.2
referencing==0.35.1
rpds-py==0.20.0
setuptools==75.1.0
sqlparse==0.5.1
typing_extensions==4.12.2
uritemplate==4.1.1
yarl==1.15.0
